+++ util.py
@@ -92,6 +92,13 @@ class Trainer:
         for epoch in range(epochs):
             train_loss = self.train_epoch(dl_train)
             val_loss = self.val_epoch(dl_val)
+
+            # Early stopping if loss becomes NaN
+            if math.isnan(train_loss) or math.isnan(val_loss):
+                print("Early stopping due to NaN loss")
+                val_loss = math.inf
+                break
+
             self.scheduler.step()
             wandb.log({
                 "train_loss": train_loss,